+++ Sources/SpotifyWebAPI/Publisher Extensions/OtherPublisherExtensions.swift
@@ -47,43 +47,6 @@ public extension Publisher where Output: Paginated {
 
 public extension Publisher {
     
-    /// Wraps all upstream elements in an optional.
-    typealias MapToNil = Publishers.Map<Self, Optional<Output>>
-    /// Replaces all errors in the stream with `nil`.
-    typealias MapErrorToNil = Publishers.ReplaceError<MapToNil>
-    
-    /// Returns a new publisher in which the output is wrapped in an optional
-    /// and errors are replaced with nil.
-    /// Therefore, the new publisher never fails.
-    func mapErrorToNil() -> MapErrorToNil {
-        
-        return self
-            .map(Optional.init)
-            .replaceError(with: nil)
-    }
-    
-    /**
-     Assigns each element from the upstream publisher to
-     an **optional** property on an object. If the upsteam publisher
-     fails with an error, then `nil` is assigned to the property.
-    
-     - Parameters:
-       - keyPath: The key path of the property to assign.
-       - object: The object on which to assign the value.
-     - Returns: A cancellable instance; used when you end assignment
-           of the received value. Deallocation of the result
-           will tear down the subscription stream.
-     */
-    func assignToOptional<Root>(
-        _ keyPath: ReferenceWritableKeyPath<Root, Optional<Self.Output>>,
-        on object: Root
-    ) -> AnyCancellable {
-        
-        return self
-            .mapErrorToNil()
-            .assign(to: keyPath, on: object)
-    }
-    
     /**
      Transforms all elements from an upstream publisher
      into a new or existing publisher.
@@ -167,7 +130,7 @@ public extension Result.Publisher where Failure == Error {
     static func catching(_ body: () throws -> Success) -> Self {
         return Self(Result(catching: body))
     }
-    
+
 }
 
 public extension Error {