+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2020-11-14
+
+### Changed
+
+* Made `Endpoints` enum Public.
+
+### Removed
+
+* Removed `Publisher.assignToOptional(_:on:)`.
+
 ## [1.0.4] - 2020-11-12
 
 ### Added